# perf: detect & fastpath no-op range slices in all layout types