--- a/src/awkward/contents/bytemaskedarray.py
+++ b/src/awkward/contents/bytemaskedarray.py
@@ -413,7 +413,6 @@ class ByteMaskedArray(ByteMaskedMeta[Content], Content):
         # (i.e. slicing the full array) and shortcut to avoid noticeable python overhead
         if (
             self._backend.nplike.known_data
-            and self.length is not unknown_length
             and (start == 0 and stop == self.length)
         ):
             return self