--- a/src/awkward/contents/indexedarray.py
+++ b/src/awkward/contents/indexedarray.py
@@ -319,6 +319,15 @@ class IndexedArray(IndexedMeta[Content], Content):
             self._touch_shape(recursive=False)
             return self
 
+        # in non-typetracer mode (and if all lengths are known) we can check if the slice is a no-op
+        # (i.e. slicing the full array) and shortcut to avoid noticeable python overhead
+        if (
+            self._backend.nplike.known_data
+            and self.length is not unknown_length
+            and (start == 0 and stop == self.length)
+        ):
+            return self
+
         return IndexedArray(
             self._index[start:stop], self._content, parameters=self._parameters
         )