--- a/src/awkward/contents/indexedoptionarray.py
+++ b/src/awkward/contents/indexedoptionarray.py
@@ -344,10 +344,7 @@ class IndexedOptionArray(IndexedOptionMeta[Content], Content):
 
         # in non-typetracer mode (and if all lengths are known) we can check if the slice is a no-op
         # (i.e. slicing the full array) and shortcut to avoid noticeable python overhead
-        if (
-            self._backend.nplike.known_data
-            and (start == 0 and stop == self.length)
-        ):
+        if self._backend.nplike.known_data and (start == 0 and stop == self.length):
             return self
 
         return IndexedOptionArray(