--- a/src/awkward/contents/listarray.py
+++ b/src/awkward/contents/listarray.py
@@ -358,10 +358,7 @@ class ListArray(ListMeta[Content], Content):
 
         # in non-typetracer mode (and if all lengths are known) we can check if the slice is a no-op
         # (i.e. slicing the full array) and shortcut to avoid noticeable python overhead
-        if (
-            self._backend.nplike.known_data
-            and (start == 0 and stop == self.length)
-        ):
+        if self._backend.nplike.known_data and (start == 0 and stop == self.length):
             return self
 
         return ListArray(