--- a/src/awkward/contents/listoffsetarray.py
+++ b/src/awkward/contents/listoffsetarray.py
@@ -345,7 +345,6 @@ class ListOffsetArray(ListOffsetMeta[Content], Content):
         # (i.e. slicing the full array) and shortcut to avoid noticeable python overhead
         if (
             self._backend.nplike.known_data
-            and self.length is not unknown_length
             and (start == 0 and stop == self.length)
         ):
             return self