--- a/src/awkward/contents/recordarray.py
+++ b/src/awkward/contents/recordarray.py
@@ -452,9 +452,14 @@ class RecordArray(RecordMeta[Content], Content):
             slice(start, stop), self.length
         )
 
-        # in non-typetracer mode we can check if the slice is a no-op
+        # in non-typetracer mode (and if all lengths are known) we can check if the slice is a no-op
         # (i.e. slicing the full array) and shortcut to avoid noticeable python overhead
-        if self._backend.nplike.known_data and (start == 0 and stop == self.length):
+        if (
+            self._backend.nplike.known_data
+            and length is not unknown_length
+            and self.length is not unknown_length
+            and (start == 0 and stop == length == self.length)
+        ):
             return self
 
         if len(self._contents) == 0: