--- a/src/awkward/contents/recordarray.py
+++ b/src/awkward/contents/recordarray.py
@@ -456,8 +456,6 @@ class RecordArray(RecordMeta[Content], Content):
         # (i.e. slicing the full array) and shortcut to avoid noticeable python overhead
         if (
             self._backend.nplike.known_data
-            and length is not unknown_length
-            and self.length is not unknown_length
             and (start == 0 and stop == length == self.length)
         ):
             return self