--- a/src/awkward/contents/recordarray.py
+++ b/src/awkward/contents/recordarray.py
@@ -454,9 +454,8 @@ class RecordArray(RecordMeta[Content], Content):
 
         # in non-typetracer mode (and if all lengths are known) we can check if the slice is a no-op
         # (i.e. slicing the full array) and shortcut to avoid noticeable python overhead
-        if (
-            self._backend.nplike.known_data
-            and (start == 0 and stop == length == self.length)
+        if self._backend.nplike.known_data and (
+            start == 0 and stop == length == self.length
         ):
             return self