--- a/src/awkward/contents/unmaskedarray.py
+++ b/src/awkward/contents/unmaskedarray.py
@@ -256,6 +256,15 @@ class UnmaskedArray(UnmaskedMeta[Content], Content):
             self._touch_shape(recursive=False)
             return self
 
+        # in non-typetracer mode (and if all lengths are known) we can check if the slice is a no-op
+        # (i.e. slicing the full array) and shortcut to avoid noticeable python overhead
+        if (
+            self._backend.nplike.known_data
+            and self.length is not unknown_length
+            and (start == 0 and stop == self.length)
+        ):
+            return self
+
         return UnmaskedArray(
             self._content._getitem_range(start, stop),
             parameters=self._parameters,