@pfackeldey 2025-09-08
@ianna I could add this fast path also to `ListOffsetArray`. This would bring it down even further from `474 μs` to `331 μs`. The main improvement is however here in the `RecordArray` layout type as that one can become pretty huge (many fields). What do you think about adding this to `ListOffsetArray` as well?